+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/analyses/analysis-forms-collection.js
@@ -13,7 +13,7 @@ var PointInPolygonFormModel = require('./analysis-form-models/point-in-polygon-f
 var UnknownTypeFormModel = require('./analysis-form-models/unknown-type-form-model');
 var WeightedCentroidModelFormModel = require('./analysis-form-models/weighted-centroid-form-model');
 var SamplingFormModel = require('./analysis-form-models/sampling-form-model');
-var SpatialMarkovTrend = require('./analysis-form-models/spatial-markov-trend');
+var SpatialMarkovTrend = require('./analysis-form-models/spatial-markov-trend-model');
 
 var TYPE_TO_ANALYSIS_FORM_MODELS_MAP = {
   'aggregate-intersection': IntersectionFormModel,
